+++ cloud/objstore/objval/bucket.go
@@ -5,3 +5,9 @@ type BucketLockingStatus struct {
 	// Enabled - if set to true then object locking must be enabled for the bucket.
 	Enabled bool
 }
+
+// BucketVersioningStatus represents the bucket-level object versioning metadata.
+type BucketVersioningStatus struct {
+	// Enabled - if set to true then object versioning must be enabled for the bucket.
+	Enabled bool
+}